+++ src/libtools/reducedgraph.cc
@@ -41,10 +41,10 @@ class GraphNode;
  * a vertex along an edge. The vertices appear in the chain in the same order
  * as they exist in a graph.
  **/
-bool compareChainWithChains_(const vector<int>& chain,
-                             const vector<vector<int>>& chains) {
+bool compareChainWithChains_(const vector<long int>& chain,
+                             const vector<vector<long int>>& chains) {
   bool match = false;
-  for (const vector<int>& chain2 : chains) {
+  for (const vector<long int>& chain2 : chains) {
     if (chain2.size() == chain.size()) {
       match = true;
       for (size_t index = 0; index < chain.size(); ++index) {
@@ -91,18 +91,18 @@ bool compareChainWithChains_(const vector<int>& chain,
  * @param[in,out] - vector of reduced edges
  * @return - set of integers containing junctions
  **/
-set<int> getVertexJunctions_(const vector<ReducedEdge>& reduced_edges) {
-  unordered_map<int, int> vertex_count;
+set<long int> getVertexJunctions_(const vector<ReducedEdge>& reduced_edges) {
+  unordered_map<long int, long int> vertex_count;
   for (ReducedEdge reduced_edge : reduced_edges) {
     // if loop, increment value is double and the first index is skipped to
     // prevent over counting of the first number
-    int increment = 1;
+    long int increment = 1;
     size_t index = 0;
     if (reduced_edge.loop()) {
       ++index;
       increment = 2;
     }
-    vector<int> chain = reduced_edge.getChain();
+    vector<long int> chain = reduced_edge.getChain();
     for (; index < chain.size(); ++index) {
       if (vertex_count.count(chain.at(index))) {
         vertex_count[chain.at(index)] += increment;
@@ -111,8 +111,8 @@ set<int> getVertexJunctions_(const vector<ReducedEdge>& reduced_edges) {
       }
     }
   }
-  set<int> junctions;
-  for (pair<int, int> vertex_and_count : vertex_count) {
+  set<long int> junctions;
+  for (pair<long int, long int> vertex_and_count : vertex_count) {
     if (vertex_and_count.second > 2) {
       junctions.insert(vertex_and_count.first);
     }
@@ -150,7 +150,7 @@ set<int> getVertexJunctions_(const vector<ReducedEdge>& reduced_edges) {
  **/
 void addEdgeIfNotLoop_(
     vector<Edge>& edges, const ReducedEdge reduced_edge,
-    unordered_map<Edge, vector<vector<int>>>& expanded_edges) {
+    unordered_map<Edge, vector<vector<long int>>>& expanded_edges) {
 
   Edge edge(reduced_edge.getEndPoint1(), reduced_edge.getEndPoint2());
   if (expanded_edges.count(edge)) {
@@ -189,7 +189,7 @@ void addEdgeIfNotLoop_(
  *
  * @param[in,out] - vector of integers containing the chain
  **/
-void orderChainAfterInitialVertex_(vector<int>& chain) {
+void orderChainAfterInitialVertex_(vector<long int>& chain) {
   size_t ignore_first_and_last_vertex = 2;
   size_t total_number_to_parse =
       (chain.size() - ignore_first_and_last_vertex) / 2;
@@ -208,17 +208,17 @@ void orderChainAfterInitialVertex_(vector<int>& chain) {
 
 bool reorderAndStoreChainIfDoesNotExist_(
     vector<Edge>& edges,
-    unordered_map<Edge, vector<vector<int>>>& expanded_edges, vector<int> chain,
-    int vertex, size_t& chain_index) {
+    unordered_map<Edge, vector<vector<long int>>>& expanded_edges,
+    vector<long int> chain, long int vertex, size_t& chain_index) {
 
   Edge edge(vertex, vertex);
   edges.push_back(edge);
-  vector<int> new_chain;
+  vector<long int> new_chain;
   for (size_t index = 0; index < chain.size(); ++index) {
     if (((chain_index + index) % chain.size()) == 0) {
       ++chain_index;
     }
-    int new_chain_index = (chain_index + index) % chain.size();
+    long int new_chain_index = (chain_index + index) % chain.size();
     new_chain.push_back(chain.at(new_chain_index));
   }
   // Ensure that the new_chain is sorted so after the first vertex they are
@@ -232,23 +232,23 @@ bool reorderAndStoreChainIfDoesNotExist_(
   return false;
 }
 
-set<int> getAllVertices_(const std::vector<ReducedEdge>& reduced_edges) {
-  set<int> vertices;
+set<long int> getAllVertices_(const std::vector<ReducedEdge>& reduced_edges) {
+  set<long int> vertices;
   for (const ReducedEdge& reduced_edge : reduced_edges) {
-    vector<int> chain = reduced_edge.getChain();
-    for (const int vertex : chain) {
+    vector<long int> chain = reduced_edge.getChain();
+    for (const long int vertex : chain) {
       vertices.insert(vertex);
     }
   }
   return vertices;
 }
 
-set<int> getAllConnectedVertices_(
-    const unordered_map<Edge, vector<vector<int>>>& expanded_edges) {
-  set<int> all_vertices;
+set<long int> getAllConnectedVertices_(
+    const unordered_map<Edge, vector<vector<long int>>>& expanded_edges) {
+  set<long int> all_vertices;
 
   for (const auto& edge_and_chains : expanded_edges) {
-    for (vector<int> chain : edge_and_chains.second) {
+    for (vector<long int> chain : edge_and_chains.second) {
       all_vertices.insert(chain.begin(), chain.end());
     }
   }
@@ -259,7 +259,7 @@ set<int> getAllConnectedVertices_(
  ******************************************************************************/
 
 void ReducedGraph::init_(vector<ReducedEdge> reduced_edges,
-                         unordered_map<int, GraphNode> nodes) {
+                         unordered_map<long int, GraphNode> nodes) {
   vector<Edge> edges;
   nodes_ = nodes;
 
@@ -269,10 +269,10 @@ void ReducedGraph::init_(vector<ReducedEdge> reduced_edges,
 
     if (reduced_edge.loop() &&
         junctions_.count(reduced_edge.getEndPoint1()) == 0) {
-      vector<int> chain = reduced_edge.getChain();
+      vector<long int> chain = reduced_edge.getChain();
       size_t chain_index = 0;
       bool edge_added = false;
-      for (int vertex : chain) {
+      for (long int vertex : chain) {
         if (junctions_.count(vertex)) {
           edge_added = reorderAndStoreChainIfDoesNotExist_(
               edges, expanded_edges_, chain, vertex, chain_index);
@@ -299,9 +299,9 @@ void ReducedGraph::init_(vector<ReducedEdge> reduced_edges,
 
 ReducedGraph::ReducedGraph(std::vector<ReducedEdge> reduced_edges) {
 
-  set<int> vertices = getAllVertices_(reduced_edges);
-  unordered_map<int, GraphNode> nodes;
-  for (const int vertex : vertices) {
+  set<long int> vertices = getAllVertices_(reduced_edges);
+  unordered_map<long int, GraphNode> nodes;
+  for (const long int vertex : vertices) {
     GraphNode gn;
     nodes[vertex] = gn;
   }
@@ -309,15 +309,15 @@ ReducedGraph::ReducedGraph(std::vector<ReducedEdge> reduced_edges) {
 }
 
 ReducedGraph::ReducedGraph(std::vector<ReducedEdge> reduced_edges,
-                           unordered_map<int, GraphNode> nodes) {
+                           unordered_map<long int, GraphNode> nodes) {
 
-  set<int> vertices = getAllVertices_(reduced_edges);
+  set<long int> vertices = getAllVertices_(reduced_edges);
   if (nodes.size() < vertices.size()) {
     throw invalid_argument(
         "The number of nodes passed into a reduced graph "
         "must be greater or equivalent to the number of vertices");
   }
-  for (const int vertex : vertices) {
+  for (const long int vertex : vertices) {
     if (nodes.count(vertex) == 0) {
       throw invalid_argument("A vertex is missing its corresponding node.");
     }
@@ -325,7 +325,7 @@ ReducedGraph::ReducedGraph(std::vector<ReducedEdge> reduced_edges,
   init_(reduced_edges, nodes);
 
   // Add all the nodes that are isolated
-  for (pair<const int, GraphNode>& id_and_node : nodes) {
+  for (pair<const long int, GraphNode>& id_and_node : nodes) {
     if (vertices.count(id_and_node.first) == 0) {
       edge_container_.addVertex(id_and_node.first);
     }
@@ -334,7 +334,7 @@ ReducedGraph::ReducedGraph(std::vector<ReducedEdge> reduced_edges,
 
 Graph ReducedGraph::expandGraph() const {
   vector<Edge> all_expanded_edges;
-  for (const pair<Edge, vector<vector<int>>> edge_and_vertices :
+  for (const pair<Edge, vector<vector<long int>>> edge_and_vertices :
        expanded_edges_) {
     vector<vector<Edge>> edges_local = expandEdge(edge_and_vertices.first);
     for (vector<Edge>& edges : edges_local) {
@@ -347,8 +347,8 @@ Graph ReducedGraph::expandGraph() const {
 
 vector<vector<Edge>> ReducedGraph::expandEdge(const Edge& edge) const {
   vector<vector<Edge>> all_edges;
-  vector<vector<int>> chains = expanded_edges_.at(edge);
-  for (vector<int> vertices : chains) {
+  vector<vector<long int>> chains = expanded_edges_.at(edge);
+  for (vector<long int> vertices : chains) {
     vector<Edge> edges;
     for (size_t index = 0; index < (vertices.size() - 1); ++index) {
       Edge edge_temp(vertices.at(index), vertices.at(index + 1));
@@ -359,17 +359,18 @@ vector<vector<Edge>> ReducedGraph::expandEdge(const Edge& edge) const {
   return all_edges;
 }
 
-vector<pair<int, GraphNode>> ReducedGraph::getNodes() const {
-  vector<int> vertices = edge_container_.getVertices();
-  vector<pair<int, GraphNode>> nodes;
-  for (const int vertex : vertices) {
-    pair<int, GraphNode> id_and_node(vertex, nodes_.at(vertex));
+vector<pair<long int, GraphNode>> ReducedGraph::getNodes() const {
+  vector<long int> vertices = edge_container_.getVertices();
+  vector<pair<long int, GraphNode>> nodes;
+  for (const long int vertex : vertices) {
+    pair<long int, GraphNode> id_and_node(vertex, nodes_.at(vertex));
     nodes.push_back(id_and_node);
   }
 
-  set<int> all_connected_vertices = getAllConnectedVertices_(expanded_edges_);
+  set<long int> all_connected_vertices =
+      getAllConnectedVertices_(expanded_edges_);
   // Grab the nodes that are not attached to any edges
-  for (pair<int, GraphNode> id_and_node : nodes_) {
+  for (pair<long int, GraphNode> id_and_node : nodes_) {
     if (!all_connected_vertices.count(id_and_node.first)) {
       nodes.push_back(id_and_node);
     }
@@ -377,11 +378,12 @@ vector<pair<int, GraphNode>> ReducedGraph::getNodes() const {
   return nodes;
 }
 
-vector<int> ReducedGraph::getVerticesDegree(int degree) const {
+vector<long int> ReducedGraph::getVerticesDegree(long int degree) const {
   if (degree == 0) {
-    set<int> all_connected_vertices = getAllConnectedVertices_(expanded_edges_);
-    vector<int> vertices;
-    for (const pair<int, GraphNode> id_and_node : nodes_) {
+    set<long int> all_connected_vertices =
+        getAllConnectedVertices_(expanded_edges_);
+    vector<long int> vertices;
+    for (const pair<long int, GraphNode> id_and_node : nodes_) {
       if (all_connected_vertices.count(id_and_node.first) == false) {
         vertices.push_back(id_and_node.first);
       }
@@ -393,7 +395,7 @@ vector<int> ReducedGraph::getVerticesDegree(int degree) const {
 
 ostream& operator<<(ostream& os, const ReducedGraph graph) {
   os << "Graph" << endl;
-  for (const pair<int, GraphNode>& id_and_node : graph.nodes_) {
+  for (const pair<long int, GraphNode>& id_and_node : graph.nodes_) {
     os << "Node " << id_and_node.first << endl;
     os << id_and_node.second << endl;
   }
@@ -404,10 +406,10 @@ ostream& operator<<(ostream& os, const ReducedGraph graph) {
   os << endl;
   os << "Expanded Edge Chains" << endl;
 
-  for (const pair<const Edge, vector<vector<int>>>& edge_and_chains :
+  for (const pair<const Edge, vector<vector<long int>>>& edge_and_chains :
        graph.expanded_edges_) {
-    for (const vector<int>& chain : edge_and_chains.second) {
-      for (const int vertex : chain) {
+    for (const vector<long int>& chain : edge_and_chains.second) {
+      for (const long int vertex : chain) {
         os << vertex << " ";
       }
       os << endl;
